move product schema to databse folder
